requirements.txt
playwright==1.41.0
gspread==6.0.2
google-auth==2.28.0
python-telegram-bot==20.7
python-dotenv==1.0.1